backend/requirements.txt
flask
flask-cors
mysql-connector-python
opencv-python
tensorflow
numpy
pillow
requests
google-generativeai
openai
